refactor(runtime): extract turn/events.rs into agent/turn — #7415
