+++ crates/zeroclaw-runtime/src/agent/loop_.rs
@@ -146,8 +146,7 @@ pub use super::cost::{
     check_tool_loop_budget, record_tool_loop_cost_usage,
 };
 
-/// Minimum characters per chunk when relaying LLM text to a streaming draft.
-const STREAM_CHUNK_MIN_CHARS: usize = 80;
+pub(crate) use super::turn::STREAM_CHUNK_MIN_CHARS;
 /// Maximum malformed internal tool-protocol retries before returning a safe fallback.
 const MAX_MALFORMED_TOOL_PROTOCOL_RETRIES: usize = 2;
 
@@ -426,24 +425,7 @@ fn compute_excluded_mcp_tools(
 
 pub use super::turn::scrub_credentials;
 
-/// Default trigger for auto-compaction when non-system message count exceeds this threshold.
-/// Prefer passing the config-driven value via `run_tool_call_loop`; this constant is only
-/// used when callers omit the parameter.
-/// Minimum interval between progress sends to avoid flooding the draft channel.
-pub const PROGRESS_MIN_INTERVAL_MS: u64 = 500;
-
-/// Delta sent from the agent loop to the channel's draft updater.
-/// Append-only — no clear/reset variant exists by design.
-#[derive(Debug, Clone)]
-pub enum StreamDelta {
-    /// Response text to append to the message buffer.
-    Text(String),
-    /// Ephemeral tool progress (not part of the response body).
-    Status(String),
-}
-
-/// Backwards-compatible alias while callers are migrated.
-pub type DraftEvent = StreamDelta;
+pub use super::turn::{DraftEvent, PROGRESS_MIN_INTERVAL_MS, StreamDelta};
 
 pub use zeroclaw_api::TOOL_CHOICE_OVERRIDE;
