+++ crates/zeroclaw-runtime/src/agent/turn/events.rs
@@ -1,0 +1,23 @@
+//! Stream/draft event types and pacing constants for the turn loop.
+
+/// Minimum characters per chunk when relaying LLM text to a streaming draft.
+pub(crate) const STREAM_CHUNK_MIN_CHARS: usize = 80;
+
+/// Default trigger for auto-compaction when non-system message count exceeds this threshold.
+/// Prefer passing the config-driven value via `run_tool_call_loop`; this constant is only
+/// used when callers omit the parameter.
+/// Minimum interval between progress sends to avoid flooding the draft channel.
+pub const PROGRESS_MIN_INTERVAL_MS: u64 = 500;
+
+/// Delta sent from the agent loop to the channel's draft updater.
+/// Append-only — no clear/reset variant exists by design.
+#[derive(Debug, Clone)]
+pub enum StreamDelta {
+    /// Response text to append to the message buffer.
+    Text(String),
+    /// Ephemeral tool progress (not part of the response body).
+    Status(String),
+}
+
+/// Backwards-compatible alias while callers are migrated.
+pub type DraftEvent = StreamDelta;
